+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add optional `stopOnError` parameter to `Lexer.lexAll`.
   - Change `Token` and `Definition` from type to interface.
   - If token has no error, the error field will be `undefined`.
+  - Add interface `ILexer`.
 - Parser
   - Use `DefinitionContextBuilder` to define parser actions.
   - If ASTNode has no error, the error field will be `undefined`.